Restructure remaining src/ files and correct respective import statements.
