+++ src/resources/mesh/objParser.ts
