+++ src/resources/textures.ts
