+++ src/runtime/renderer.ts
@@ -10,12 +10,12 @@
  * - Comprehensive error handling and reporting
  */
 
-import { parseWGSL, parseErrorMessages } from './parser';
-import { loadDefaultTexture, usesAnyTextures } from './textures';
-import { getWebGPUDevice, configureCanvasContext } from './context';
-import { fullscreenVertexWGSL, injectedHeader, minimalHeader, compileShaderModule } from './shaderUtils';
-import { createUniforms } from './uniforms';
-import { createPipeline } from './pipeline';
+import { parseWGSL, parseErrorMessages } from '../shader/wgslReflect';
+import { loadDefaultTexture, usesAnyTextures } from '../resources/textures';
+import { getWebGPUDevice, configureCanvasContext } from '../core/context';
+import { fullscreenVertexWGSL, injectedHeader, minimalHeader, compileShaderModule } from '../shader/shaderUtils';
+import { createUniforms } from '../pipeline/uniforms';
+import { createPipeline } from '../pipeline/pipeline';
 
 /** Current animation frame ID for render loop cancellation */
 let currentFrameId: number | null = null;
