+++ tests/backend/parser.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { parseWGSL, parseErrorMessages } from '@/webgpu/parser'
+import { parseWGSL, parseErrorMessages } from '@/shader/wgslReflect'
 
 // Minimal valid fragment shader
 const validFragmentOnly = `
